requirements.txt
pytesseract~=0.3.13
Flask~=3.1.1
Werkzeug~=3.1.3
pillow~=11.3.0